added validation for member routes

+++ AppOneDrive/app/Http/Controllers/MemberController.php
@@ -7,6 +7,8 @@ use App\Models\Member;
 use DateTime;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Validator;
+use Illuminate\Validation\Rule;
 
 use function PHPUnit\Framework\isNull;
 
@@ -33,17 +35,50 @@ class MemberController extends Controller
      */
     public function store(Request $request)
     {
+
+        $validator = Validator::make($request->all(), [
+            'user_id' => 'required|numeric',
+            'firm_pib' => 'required|string|size:9',
+            'privileges' => [
+                'required',
+                'string',
+                'max:255',
+                Rule::in(['Read', 'Write']),
+            ]
+            ,
+        ]);
+ 
+        if($validator->fails()){
+            return response()->json($validator->errors());
+        }
+
         $user_id=$request->input('user_id');
         $firm_id=$request->input('firm_pib');
         $privileges=$request->input('privileges');
-        $member=Member::create(['user_id'=>$user_id,'firm_pib'=>$firm_id,'AddedAt'=>new DateTime(),'privileges'=>$privileges]);
-        return response()->json(['message' => 'Member added successfully', 'member' => $member], 201);
+
+        try{
+            $member=Member::create(['user_id'=>$user_id,'firm_pib'=>$firm_id,'AddedAt'=>new DateTime(),'privileges'=>$privileges]);
+            return response()->json(['message' => 'Member added successfully', 'member' => $member], 201);
+        }
+        catch(\Exception $ex){
+            return response()->json(['message' => 'Error while adding user to firm'], 400);
+        }
+        
     }
 
     // Gets a specific member
     public function show(Request $request)
     {
-        $mem = Member::where('firma_pib', $request->route("PIB"))->get();
+        $pib = $request->route('PIB');
+        $validator = Validator::make(['PIB' => $pib], [
+            'PIB' => 'required|string|size:9',
+        ]);
+        if($validator->fails()){
+            return response()->json($validator->errors(),400);
+        }
+
+        $mem = Member::where('firm_pib', $pib)->get();
+        if($mem->count()==0)return response()->json(["message"=>"Firm not found"],404);
         return MemberResource::collection($mem);
     }
 
@@ -60,8 +95,24 @@ class MemberController extends Controller
      */
     public function update(Request $request, $ID,$PIB)
     {
+        $validator = Validator::make(['ID'=>$ID,'PIB' => $PIB,'privileges'=>$request->input('privileges')], [
+            'ID'=>'required|numeric',
+            'PIB' => 'required|string|size:9',
+            'privileges' => [
+                'required',
+                'string',
+                'max:255',
+                Rule::in(['Read', 'Write']),
+            ]
+        ]);
+
+        if($validator->fails()){
+            return response()->json($validator->errors(),400);
+        }
+
         $result=Member::Where('user_id',$ID)->Where('firm_pib',$PIB)->update(['privileges'=>$request->input('privileges')]);
         if($result==1)return response()->json(['message' => 'Member updated successfully', 'member' => Member::Where('user_id',$ID)->Where('firm_pib',$PIB)->get()], 201);
+        else return response()->json(['message' => 'Error while updating member or there is nothing to update'],400);
     }
 
     /**
@@ -69,7 +120,17 @@ class MemberController extends Controller
      */
     public function destroy($ID,$PIB)
     {
+        $validator = Validator::make(['ID'=>$ID,'PIB' => $PIB], [
+            'ID'=>'required|numeric',
+            'PIB' => 'required|string|size:9',
+        ]);
+
+        if($validator->fails()){
+            return response()->json($validator->errors(),400);
+        }
+
         $result=Member::Where('user_id',$ID)->Where('firm_pib',$PIB)->delete();
         if($result==1) return response()->json(['message' => 'Member deleted successfully'], 201);
+        else return response()->json(['message' => 'Error while deleting member'],400);
     }
 }
